### 2019-11-21 #1908 @whedon
[PRE REVIEW]: profileR: An R package for profile analysis
**Submitting author:** @cddesja (<a href="http://orcid.org/0000-0002-8234-9665">Christopher Desjardins</a>)
**Repository:** <a href="https://github.com/cddesja/profileR" target ="_blank">https://github.com/cddesja/profileR</a>
**Version:** 0.3-5
