### 2019-11-21 #1908 @kthyng
Hi @cddesja and thanks for your submission! Our system hasn't found an open source license and that is a requirement for being part of JOSS. Do you want to add one?
