### 2019-11-21 #1908 @kthyng
Hi @marcosvital! What do you think about this being your first JOSS submission as editor?
