### 2019-11-21 #1908 @cddesja
Hi,

Sorry, yes. The software is GPL v3. Where should I add this?

On Thu, Nov 21, 2019 at 5:31 PM Kristen Thyng <notifications@github.com>
wrote:

> Hi @marcosvital <https://github.com/marcosvital>! What do you think about
> this being your first JOSS submission as editor?
>
