### 2019-11-22 #1908 @markhwhiteii
I'm not an expert in this type of statistical analysis, but it seems related enough to work I've done that I could review. This would be my first time reviewing—is there a guide on the typical type of review process?
