+++ man/writing2.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{writing2}
+\alias{writing2}
+\title{Second hypothetical writing prompt example}
+\format{A data frame with 100 rows and 4 variables:
+\describe{
+  \item{student}{The student identifier}
+  \item{prompt}{The writing prompt identifier. There were five prompts}
+  \item{rater}{The rater identifier. There were two raters.}
+  \item{score}{Student score on the writing prompt. Ranged from 4 - 100, with a maximum score of 100.}
+}}
+\source{
+Desjardins, C. D. & Bulut, O. (2017). Using R for Educational and Psychological Measurement.
+}
+\usage{
+writing2
+}
+\description{
+This is a hypothetical data set to demonstrate a partially nested two-facet design for generalizability theory. In this design, the raters are nested within students, W X (S:R). See chapter 3 of "Using R for Educational and Psychological Measurement" for more details.
+}
+\examples{
+library(lme4)
+nested.design <- lmer(scores ~ (1 | students/raters) + (1 | prompts) + (1 | students:prompts), data = writing2)
+gstudy(nested.design)
+}
+\keyword{dataset}
