### 2019-11-22 #1908 @jrosen48
Hi, I'd be happy to review this. I have experience with this type of analysis (and developing [software](https://github.com/data-edu/tidyLPA) for it) and think/hope I'd be able to contribute.
