### 2019-11-22 #1908 @kthyng
> Hi, Sorry, yes. The software is GPL v3. Where should I add this?

@cddesja See details about the license here: https://joss.readthedocs.io/en/latest/review_criteria.html#software-license
