### 2019-11-26 #1908 @wjakethompson
I would also be happy to review. I haven't worked with these models specifically, but I've done some work on closely related models, also in R.
