### 2019-11-27 #1908 @marcosvital
Hi, @cddesja , I'm Marcos, and I'll be the editor of your submission.

Since we already have three reviewers, we are almost ready to move from pre-review to review, so please confirm if everything is okay with the article proof and we'll move on.

@wjakethompson , @jrosen48 and @markhwhiteii , thank you all for accepting review this submission. You will all get the detailed reviewing instructions when we move to the next phase, and please fell free to call me anytime if you have any questions.
