### 2019-12-02 #1908 @cddesja
Here's a couple strange things in the PDF.

- It looks like the paragraph below the first Figure no longer says Figure
1.
- Right before the second Figure it should say (see Figure 2).

On Wed, Nov 27, 2019 at 2:39 PM Marcos Vital (LEQ-UFAL) <
notifications@github.com> wrote:

> Hi, @cddesja <https://github.com/cddesja> , I'm Marcos, and I'll be the
> editor of your submission.
>
> Since we already have three reviewers, we are almost ready to move from
> pre-review to review, so please confirm if everything is okay with the
> article proof and we'll move on.
>
> @wjakethompson <https://github.com/wjakethompson> , @jrosen48
> <https://github.com/jrosen48> and @markhwhiteii
> <https://github.com/markhwhiteii> , thank you all for accepting review
> this submission. You will all get the detailed reviewing instructions when
> we move to the next phase, and please fell free to call me anytime if you
> have any questions.
>
