### 2019-12-09 #1908 @marcosvital
Hi, @cddesja , and sorry for the delayed reply.

The problem with the figure captions is probably easy to solve: you should try to insert the figure in a single separated line in the md file, with the caption inside brackets.

Take a look at this example: https://raw.githubusercontent.com/pyvista/pyvista/master/joss/paper.md

That will generate this: https://github.com/pyvista/pyvista/blob/master/joss/paper.md
