### 2019-12-09 #1908 @marcosvital
Reviewers, as soon as the proof gets ready, we are ready to begin with the reviewing process. Please take a look at the guidelines if it's your first time reviewing for JOSS:

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
