### 2019-12-10 #1908 @cddesja
@marcosvital <https://github.com/marcosvital>, the figures should now be
fixed and the proof should be ready.

On Mon, Dec 9, 2019 at 7:59 AM Marcos Vital (LEQ-UFAL) <
notifications@github.com> wrote:

> Reviewers, as soon as the proof gets ready, we are ready to begin with the
> reviewing process. Please take a look at the guidelines if it's your first
> time reviewing for JOSS:
>
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
