### 2019-12-10 af2b69a
Update paper.md
