### 2019-12-10 #1908 @marcosvital
@cddesja , I think that everything is fine with the proof now. Please take a look, and we can jump to review as soon as you say we are ready.
