### 2019-12-10 #1908 @cddesja
Can you recompile it? I updated the text for Figure 2 and forgot to push it
to GitHub and itshould be current there now.

On Tue, Dec 10, 2019 at 12:34 PM Marcos Vital (LEQ-UFAL) <
notifications@github.com> wrote:

> @cddesja <https://github.com/cddesja> , I think that everything is fine
> with the proof now. Please take a look, and we can jump to review as soon
> as you say we are ready.
>
