### 2019-12-10 #1941 @whedon
[REVIEW]: profileR: An R package for profile analysis
**Submitting author:** @cddesja (<a href="http://orcid.org/0000-0002-8234-9665">Christopher Desjardins</a>)
**Repository:** <a href="https://github.com/cddesja/profileR" target ="_blank">https://github.com/cddesja/profileR</a>
**Version:** v0.3-6

@wjakethompson & @jrosen48 & @markhwhiteii , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
