### 2019-12-10 509412c
removed dstudy and gstudy because they are presently broken.
