+++ man/print.dStudy.Rd
@@ -1,19 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dstudy.R
-\name{print.dStudy}
-\alias{print.dStudy}
-\title{Prints the results from the D-study}
-\usage{
-\method{print}{dStudy}(x, ...)
-}
-\arguments{
-\item{x}{object of class \code{dStudy} returned from the \code{dstudy} function}
-
-\item{...}{additional objects of the same type.}
-}
-\description{
-Prints  output from fitting the \code{dstudy} function.
-}
-\seealso{
-\code{\link{dstudy}}
-}
