### 2020-02-08 #1941 @jrosen48
Hi @marcosvital, I seem to be unable to check the checkbox items - can you advise if there is something I can do? thanks!
