### 2020-02-08 cddesja/profileR#27 @jrosen48
Add CRAN installation instructions
This pertains to [JOSS review 1941](https://github.com/openjournals/joss-reviews/issues/1941).

Because the README includes code to install the developmental version on GitHub, readers may not think that a version is available through CRAN (though I note that the README does mention this). What do you think of including `install.packages('profileR')` to the README?

Relatedly, what do you think of including installation instructions in the paper?
