### 2020-02-08 cddesja/profileR#28 @jrosen48
Improve example usage
This pertains to [JOSS review 1941](https://github.com/openjournals/joss-reviews/issues/1941).

Example usage is not documented in the README or paper; there is also not a vignette. Because of this, it may be somewhat challenging for those new to the package to quickly understand how to use the package, and, specifically, because the documentation is primarily specific to each function, how the functions can be used together to carry out a profile analysis.
