### 2020-02-08 cddesja/profileR#29 @jrosen48
Include tests
This pertains to [JOSS review 1941](https://github.com/openjournals/joss-reviews/issues/1941).

The review criteria require that the following be addressed; I do not yet see these in the repository or documentation:

`Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`
