### 2020-02-08 cddesja/profileR#30 @jrosen48
Include community guidelines
This pertains to [JOSS review 1941](https://github.com/openjournals/joss-reviews/issues/1941).

The review criteria require that the following be addressed; I do not yet see these in the repository or documentation:

`Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`
