### 2020-02-08 cddesja/profileR#31 @jrosen48
Include a review of the state of the field
This pertains to [JOSS review 1941](https://github.com/openjournals/joss-reviews/issues/1941).

The review criteria require that the following be addressed; I do not yet see these in the repository or documentation:

`State of the field: Do the authors describe how this software compares to other commonly-used packages?`

This is selfish, but could a statement of how this package differs from another on profile analysis associated with JOSS ([tidyLPA](https://joss.theoj.org/papers/10.21105/joss.00978)). To be sure, I can see that this package includes functionality above and beyond what tidyLPA offers, and I think there is room for more than one package for carrying out profile analysis where it comes to JOSS and R (of course), but think that a distinction between the two may be useful to readers.
