### 2020-02-08 cddesja/profileR#27 @cddesja
I think both of these are good ideas. I'll update the README.md file for that and make sure that the installations instruction get added to the paper.
