### 2020-02-08 cddesja/profileR#28 @cddesja
@okanbulut and I have an unpublished vignette that we're going to submit to https://psyarxiv.com/. @jrosen48, would you like to see an actual example in the paper or would linking to the article there be sufficient?
