### 2020-02-08 cddesja/profileR#29 @cddesja
@okanbulut and I will need to think about this. We don't have any tests add. We can look at the testthat package. @jrosen48, is this something that should have been added before the article was submitted?
