### 2020-02-08 cddesja/profileR#31 @cddesja
@jrosen48, sure. We can add a couple of sentences to the manuscript that compare our package with yours and our unpublished vignette has a greater comparison of these approaches.

@jrosen48, do you happen to know if we're suppose to fix the manuscript based on these issues, close the issue, and then inform you and the editor?
