### 2020-02-08 #1941 @kthyng
Hi @jrosen48 — did you receive an invitation to edit by email? You need to accept that to gain proper permissions on this issue. Let me know if you don't have the invitation.
