### 2020-02-08 cddesja/profileR#30 @jrosen48
What would you think of adding this to the repository - perhaps to the README - instead of to the article? I think the article is more for general info, whereas the repository is for folks who may wish to do these things (contribute, report issues, find help/support).
