### 2020-02-08 cddesja/profileR#29 @jrosen48
Yes, I think so. It is clearly in the review criteria, and all review criteria must be met before acceptance.
