### 2020-02-08 cddesja/profileR#28 @jrosen48
I think either would work, and I would be satisfied with either of the options you proposed. My slight preference would be to see a short, worked example (perhaps for the plots that are presented - as-is, it is unclear how they are generated).

Just a question, why is the vignette not published with the package? I think this could be useful!
