### 2020-02-08 cddesja/profileR#31 @jrosen48
I think that process that you described is exactly right! In my experience (as an author and reviewer), there are usually at least a few changes to the paper needed, and when there are changes to the paper, you can have it re-rendered.
