### 2020-02-09 cddesja/profileR#32 @wjakethompson
LICENSE consistency
This pertains to openjournals/joss-reviews#1941.

The LICENSE file includes the GPL v3.0 license, but the DESCRIPTION file specifies GPL (>=2.0). Given the fairly significant differences between v2.0 and v3.0, I would pick one and make it consistent between these two locations.
