### 2020-02-09 cddesja/profileR#28 @wjakethompson
I agree with @jrosen48 that including the vignette within in the package would be helpful. For someone who just finds the package, it is probably easier to open a vignette built into the package. Otherwise, you rely on them finding the README, the JOSS paper, or the psyarxiv paper. In general, my preference is to have as much of the documentation as possible self-contained to improve usability.

If you don't want to publish the vignette in both places, I would at least consider adding a simplified version to the package. I would also recommend including a very minimal example in the README, but would also be fine with a link in the README to a vignette.
