### 2020-02-09 cddesja/profileR#30 @wjakethompson
I agree that the repository is probably a better place for this to live. An alternative to the README would be to add a CODE_OF_CONDUCT.md and CONTRIBUTING.md file. You probably don't need anything this extensive, but the [ggplot2 repository](https://github.com/tidyverse/ggplot2) has really good examples of both.
