### 2020-02-09 cddesja/profileR#33 @wjakethompson
paos printing of results
This pertains to openjournals/joss-reviews#1941.

This is super minor, but something I noticed as I was going through the example uses. For the `paos()` function, when running the example code as written, there are no problems, and everything prints as expected.

``` r
library(profileR)
#> Loading required package: ggplot2
#> Loading required package: RColorBrewer
#> Loading required package: reshape
#> Loading required package: lavaan
#> This is lavaan 0.6-5
#> lavaan is BETA software! Please report any bugs.

data(nutrient)
paos(nutrient, scale = TRUE)
#>
#> Profile Analysis for One Sample with Hotelling's T-Square:
#>                                               T-Squared        F df1 df2 p-value
#> Ho: Ratios of the means over Mu0=1             1392.347 276.9559   5 732       0
#> Ho: All of the ratios are equal to each other  1278.073 318.2159   4 733       0
```

However, if the output is saved (as it might be if you wanted to reference the results of the hypothesis test in an R Markdown document), the printing is a little odd. The heading text (`Profile Analysis for One Sample with Hotelling's T-Square:`) prints when the function executes, even though the result is not printed. Then the heading text is not printed when the results are actually printed.

```r
library(profileR)
#> Loading required package: ggplot2
#> Loading required package: RColorBrewer
#> Loading required package: reshape
#> Loading required package: lavaan
#> This is lavaan 0.6-5
#> lavaan is BETA software! Please report any bugs.

data(nutrient)
result <- paos(nutrient, scale = TRUE)
#>
#> Profile Analysis for One Sample with Hotelling's T-Square:
result
#>                                               T-Squared        F df1 df2 p-value
#> Ho: Ratios of the means over Mu0=1             1392.347 276.9559   5 732       0
#> Ho: All of the ratios are equal to each other  1278.073 318.2159   4 733       0
```

It appears the code for this is [here](https://github.com/cddesja/profileR/blob/509412cffd039cd66a9fc61905105fec4cf5144d/R/paos.R#L84-L85). If the header text is necessary, I would probably turn the result into an S3 class and create a print method. I also think it would be fine to remove the header text, as the function documentation clearly states that the Hotelling's T-square is being used.
