### 2020-02-09 #1941 @wjakethompson
I have completed my initial review. Overall, it looks like great package! A summary of why some boxes have not been checked:

### Acceptance blockers
* **Example usage**: See cddesja/profileR#28.
* **Automated tests**: See cddesja/profileR#29.
* **Community guidelines**: See cddesja/profileR#30.

### Other issues
* **License**: A license is included, conflicts with the license listed in the DESCRIPTION. See cddesja/profileR#32.
* **Installation instructions**: Instructions are included, and work. But I think CRAN instructions should also be included. See cddesja/profileR#27.
* **State of the field**: No comparisons are made to other packages. See cddesja/profileR#31.
* **References**: Assuming packages are added for state of the field, appropriate references would also need to be included.
