### 2020-02-10 cddesja/profileR#33 @okanbulut
We removed the title text from the output since it is clear what is being printed. Thanks for the suggestion.
