+++ DESCRIPTION
@@ -20,4 +20,4 @@ Depends:
     R (>= 3.0.0)
 License: GPL (>= 2)
 LazyData: true
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.2
