### 2020-02-10 cddesja/profileR#32 @okanbulut
Now we use GPL-3 in both places. Thanks for pointing that out.
