+++ DESCRIPTION
@@ -18,6 +18,6 @@ Depends:
     reshape,
     lavaan,
     R (>= 3.0.0)
-License: GPL (>= 2)
+License: GPL-3
 LazyData: true
 RoxygenNote: 7.0.2
