+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,14 @@
+
+---
+output: github_document
+---
+
+## Contributor Code of Conduct
+
+As the maintainers of ``profileR``, we pledge to respect all people who want to contribute to the package through reporting issues, posting feature requests, updating documentation, submitting pull requests or patches, and other activities. In addition, we pledge to act and interact in ways that contribute to an open, welcoming, diverse, inclusive, and healthy community.
+
+Project maintainers will retain the right and responsibility to remove, edit, or reject comments, commits, code, issues, and other contributions that are not aligned to this Code of Conduct. Instances of unacceptable behavior may be reported by opening an issue, emailing one of the core developers (Okan or Chris).
+
+This Code of Conduct is adapted from the Contributor Covenant, version 2.0, available at <https://www.contributor-covenant.org/version/2/0/code_of_conduct.html>.
+
+Community Impact Guidelines were inspired by [Mozilla’s code of conduct enforcement ladder](https://github.com/mozilla/diversity).
