### 2020-02-10 cddesja/profileR#30 @okanbulut
We have now included all of the requested information in the README document. We also included a Code of Conduct document. Thanks for the great suggestions!
