### 2020-02-10 cddesja/profileR#27 @okanbulut
We have the installation instructions in README.md. We will add a similar information in the paper as well @cddesja
