### 2020-02-10 cddesja/profileR#28 @cddesja
@okanbulut added the link to the vignette on the README.md and I'll add the vignette to the package later this week.
