### 2020-02-11 c6f7f1e
Added the vignette
