+++ README.md
@@ -42,6 +42,7 @@ install.packages("profileR")
 and the developmental version on Github can be installed by using:
 
 ```R
+Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS=TRUE) # in case your R version is older
 devtools::install_github(repo = "cddesja/profileR")
 ```
 
