### 2020-02-11 cddesja/profileR#28 @okanbulut
The vignette is now available on the development version of the package on Github.
