+++ README.md
@@ -1,5 +1,5 @@
 
-# profileR
+# profileR <img src="man/figures/logo.png" align="right" width="120" />
 
 **Development version**: v0.3-5 on GitHub [![Build Status](https://travis-ci.org/cddesja/profileR.svg?branch=master)](https://travis-ci.org/cddesja/profileR) [![codecov.io](http://codecov.io/github/cddesja/profileR/coverage.svg?branch=master)](http://codecov.io/github/cddesja/profileR?branch=master) 
 [![](https://cranlogs.r-pkg.org/badges/profileR)](https://cran.r-project.org/package=profileR)
