### 2020-02-12 478ab21
fixed LICENSE
removed LICENSE file and updated DESCRIPTION
