+++ README.md
@@ -46,7 +46,7 @@ Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS=TRUE) # in case your R version is o
 devtools::install_github(repo = "cddesja/profileR") # without the vignette
 ```
 
-If \LaTeX is installed in your computer, you can also install the package with its vignette:
+If $\LaTeX$ is installed in your computer, you can also install the package with its vignette:
 
 ```R
 Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS=TRUE) # in case your R version is older
