### 2020-02-12 cddesja/profileR#28 @cddesja
And the vignette can be accessed via the package on GitHub (and CRAN once tests are added) via:
```r
vignette("profiler-vignette")
```
