### 2020-02-12 d63e961
added some first tests and updated paper.md
