+++ DESCRIPTION
@@ -21,3 +21,5 @@ Depends:
 License: GPL-3
 LazyData: true
 RoxygenNote: 7.0.2
+Suggests: 
+    testthat (>= 2.1.0)
