+++ paper.bib
@@ -59,3 +59,24 @@
     url = {https://CRAN.R-project.org/package=psych},
   }
 
+
+@article{rosenberg2019tidylpa,
+  title={tidyLPA: An R package to easily carry out Latent Profile Analysis (LPA) using open-source or commercial software},
+  author={Rosenberg, Joshua and Beymer, Patrick and Anderson, Daniel and van Lissa, CJ and Schmidt, Jennifer},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={30},
+  pages={978},
+  year={2018}
+}
+
+@Article{mclust,
+    title = {{mclust} 5: clustering, classification and density estimation using {G}aussian finite mixture models},
+    author = {Luca Scrucca and Michael Fop and Thomas Brendan Murphy and Adrian E. Raftery},
+    journal = {The {R} Journal},
+    year = {2016},
+    volume = {8},
+    number = {1},
+    pages = {205--233},
+    url = {https://journal.r-project.org/archive/2016-1/scrucca-fop-murphy-etal.pdf},
+  }
