+++ paper.md
@@ -42,7 +42,9 @@ or for the development version on GitHub
 devtools::install_github("cddesja/profileR")
 ```
 
-Similar packages exist in R, such as ``tidyLPA`` [@rosenberg2019tidylpa] and ``mclust`` [@mclust], but our approach does not rely mixture models and is not latent. The R package ``profileR``, currently in version 0.3-6 on the comprehensive R archive network (CRAN), implements profile reliability [@bulut2013between], criterion-related profile analysis [@davison_identifying_2002], profile analysis via multidimensional scaling [@pams], moderated profile analysis, profile analysis by group, and a within-person factor model to derive score profiles [@davison2009factor] as well as a variety of graphical methods to visualize profiles. The API for ``profileR`` was designed to provide a unified, consistent, and user-friendly R interface for these methods (see Figure 2). It uses the S3 class and many generics have been written to work with ``profileR`` objects. Any future statistical methods added to ``profileR`` will adhere to the existing API. 
+ The R package ``profileR``, currently in version 0.3-6 on the comprehensive R archive network (CRAN), implements profile reliability [@bulut2013between], criterion-related profile analysis [@davison_identifying_2002], profile analysis via multidimensional scaling [@pams], moderated profile analysis, profile analysis by group, and a within-person factor model to derive score profiles [@davison2009factor] as well as a variety of graphical methods to visualize profiles. Related profile analysis packages exist in R, such as ``tidyLPA`` [@rosenberg2019tidylpa] and ``mclust`` [@mclust], however, our approach does not rely on mixture modeling and employs a wider range of profile analytic techniques.
+
+The API for ``profileR`` was designed to provide a unified, consistent, and user-friendly R interface for these methods (see Figure 2). It uses the S3 class and many generics have been written to work with ``profileR`` objects. Any future statistical methods added to ``profileR`` will adhere to the existing API (such as the addition of Bayesian approaches and multilevel profile analysis). 
 
 ![Primary functions and their generic functions available in profileR](profileR.png)
 
