+++ paper.bib
@@ -80,3 +80,10 @@
     pages = {205--233},
     url = {https://journal.r-project.org/archive/2016-1/scrucca-fop-murphy-etal.pdf},
   }
+
+@book{muthen2004mplus,
+  title={Mplus user's guide. Eighth Edition},
+  author={Muth{\'e}n, Linda K and Muth{\'e}n, Bengt O},
+  year={1998 -- 2017},
+  publisher={Muth{\'e}n \& Muth{\'e}n}
+}
