+++ paper.bib
@@ -84,6 +84,6 @@
 @book{muthen2004mplus,
   title={Mplus user's guide. Eighth Edition},
   author={Muth{\'e}n, Linda K and Muth{\'e}n, Bengt O},
-  year={1998 -- 2017},
+  year={2017},
   publisher={Muth{\'e}n \& Muth{\'e}n}
 }
