### 2020-02-12 cddesja/profileR#31 @cddesja
Added a reference to tidyLPA, mclust, and Mplus and made a distinction of our package from these methods.
