### 2020-02-12 1fa4021
Update testthat.R
