### 2020-02-12 cddesja/profileR#29 @cddesja
I think I need some clarity on what "**Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?" means exactly. We'd prefer to not use the testthat package to create tests, primarily because Wickham package's are a [moving target](https://www.quora.com/Do-you-expect-the-tidyverse-to-be-the-part-of-core-R-packages-some-day) and functions are often deprecated and packages get orphaned.

From a user's perspective, if

1. The package installs without error and
2. The examples run and can be compared, for example against a wiki.

I would say the functionality of the software can be verified.
