### 2020-02-13 cddesja/profileR#29 @jrosen48
I don't agree with your description of Wickham's packages, but, I take your point.

You can include tests without using Wickham's packages, so I think this comes down to your argument about the manual steps that can verify the functionality.

I'd suggest tests, but could be convinced that the steps you described - if elaborated and formalized a little, maybe - are satisfactory for this review criterion.
