+++ README.md
@@ -55,6 +55,10 @@ devtools::install_github(repo = "cddesja/profileR", build_vignettes = TRUE) # wi
 
 ***
 
+### Verifying and testing ``profileR``
+
+To verify the installation and test that the examples work as intended, please this our [wiki](https://github.com/cddesja/profileR/wiki/Verifying-the-installation-of-profileR) 
+
 ### Using and Citing ``profileR``
 
 To learn how to conduct profile analysis using ``profileR``, you can see our paper available on [PsyArxiv](https://psyarxiv.com/sgy8m). This paper walks through some of the principal methods available in ``profileR``. In the paper, a brief theory behind each method is presented, followed by a working example demonstrating how to use these methods in ``profileR``. To see this paper as a vignette in the package:
