### 2020-02-14 cddesja/profileR#29 @cddesja
@jrosen48, what are your thoughts about [this](https://github.com/cddesja/profileR/wiki/Verifying-the-installation-of-profileR)? We can link to it on the README.md file and I can add a note in the paper.md if you think it works. If that isn't satisfactory, please let us know.

Also, I'm not trying to denigrate Wickham's work or the greater work of the collection of packages in the tidyverse. Their stuff is amazing and Wickham is incredibly prolific. I have just had my own experiences revisiting analyses from the past that have made me want to stay as close to base R as possible.
