### 2020-02-14 cddesja/profileR#29 @jrosen48
I am satisfied with this. Idea to take or leave. What if this were in a vignette? That way, it would have to get built (i.e., upon installation and CRAN submission, in two key cases). That way, if there were ever an error introduced in one of the functions used in one of the 24 examples, you would know that way.
