### 2020-02-14 cddesja/profileR#29 @cddesja
@jrosen48, I think that's a good idea. I'll add it as a vignette.
