### 2020-02-16 f56cfa2
added a checking profiler vignette
