### 2020-02-16 #1941 @cddesja
@jrosen48 and @wjakethompson, I believe we have addressed all the outstanding issues. @jrosen48, please let us know if you think the state of the field has been described adequately.
