### 2020-02-22 #1941 @cddesja
@wjakethompson, we are wondering if there are changes are adequate for you and whether you have any additional changes?

@kthyng, we haven't heard back from @markhwhiteii. should we try and get another reviewer or if both @wjakethompson and @jrosen48 are satisfied with the state of the paper, is that sufficient?
