### 2020-03-02 #1941 @marcosvital
Hi, @kthyng , yes, I'm comfortable with moving on with two reviews, as they are both satisfied with the paper.

@wjakethompson ad @jrosen48 , thank you both so much for taking time to review and make suggestions.
@jrosen48, I noticed that there is one last box left unchecked in you review, regarding "Functionality documentation". Do you fell there is still something concerning this that the authors should address?

@cddesja , we are almost done here. Once we confirm that everything is ok to move forward, you will need to archive the last release of the package (on Zenodo, figshare, or other), if you already didn't do this, post the version number and archive DOI here in this issue.
