### 2020-03-10 #1941 @cddesja
@marcosvital, sounds good. Are you waiting on me, right now? Or are we waiting on

> @jrosen48, I noticed that there is one last box left unchecked in you review, regarding "Functionality documentation". Do you fell there is still something concerning this that the authors should address?
