### 2020-03-10 #1941 @marcosvital
Hi, @cddesja! I was waiting to know about the unchecked box. But since it's just a single item from the whole list, I'll make a final and quick check on everything myself, and we'll move on, ok? Will be back here soon.
