### 2020-03-10 #1941 @marcosvital
@cddesja, please check out the manuscript references: they should have the DOIs listed, ok?
