### 2020-03-10 #1941 @marcosvital
@markhwhiteii, as JOSS requires only two reviewers, we are proceeding with the paper acceptance. Since the reviewers are listed on the published paper, I'll remove your name from the list. I understand that taking time to review a paper is not easy, so thanks for trying to help with this one, and I hope we can count on you in future reviewing tasks.
