+++ paper.bib
@@ -2,12 +2,8 @@
 @article{davison_identifying_2002,
 	title = {Identifying criterion-related patterns of predictor scores using multiple regression.},
 	volume = {7},
-	issn = {1082-989X},
-	url = {http://doi.apa.org/getdoi.cfm?doi=10.1037/1082-989X.7.4.468},
 	doi = {10.1037//1082-989X.7.4.468},
-	language = {en},
 	number = {4},
-	urldate = {2018-04-24},
 	journal = {Psychological Methods},
 	author = {Davison, Mark L. and Davenport, Jr., Ernest C.},
 	year = {2002},
@@ -16,7 +12,7 @@
 }
 
 @incollection{pams,
-  title={Multidimensional Scaling Models of Personality Responding.},
+  title={Multidimensional scaling models of personality responding.},
   author={Davison, Mark L},
   editor={Strack, Stephen and Lorr, Maurice},
   booktitle={Differentiating Normal and Abnormal Personality},
@@ -32,20 +28,32 @@
   school={University of Minnesota}
 }
 
+@article{bulut2017,
+author = {Okan Bulut and Mark L. Davison and Michael C. Rodriguez},
+title = {Estimating Between-Person and Within-Person Subscore Reliability with Profile Analysis.},
+journal = {Multivariate Behavioral Research},
+volume = {52},
+number = {1},
+pages = {86--104},
+year  = {2017},
+doi = {10.1080/00273171.2016.1253452}}
+
 @article{davison2009factor,
   title = {Factor Analytic Modeling of Within Person Variation in Score Profiles},
   author = {Davison, Mark L and Kim, S K and Close, Catherine},
   journal = {Multivariate Behavioral Research},
   year = {2009},
   volume = {44},
-  pages ={668--687}
+  pages ={668--687},
+  doi = {10.1080/00273170903187665}
 }
 
 @book{desjardins2018handbook,
   title={Handbook of educational measurement and psychometrics using {R}},
   author={Desjardins, Christopher D and Bulut, Okan},
   year={2018},
-  publisher={Chapman and Hall/CRC}
+  publisher={Chapman and Hall/CRC},
+  doi = {10.1201/b20498}
 }
 
  @Manual{psych,
@@ -67,7 +75,8 @@
   volume={3},
   number={30},
   pages={978},
-  year={2018}
+  year={2018},
+  doi = {10.21105/joss.00978}
 }
 
 @Article{mclust,
@@ -79,6 +88,7 @@
     number = {1},
     pages = {205--233},
     url = {https://journal.r-project.org/archive/2016-1/scrucca-fop-murphy-etal.pdf},
+    doi = {10.32614/rj-2016-021}
   }
 
 @book{muthen2004mplus,
