### 2020-03-10 #1941 @okanbulut
> @cddesja, please check out the manuscript references: they should have the DOIs listed, ok?

@marcosvital, I fixed the missing DOIs in the paper. They all should be good now. Thank for catching that.
