### 2020-03-11 #1941 @cddesja
@marcosvital, I'd like to make a reference to our two vignettes included in the package by adding the following sentence bolded below. Is this okay? I'll add the citation to the paper.bib if this is fine.

The R package ``profileR``, currently in version 0.3-6 on the comprehensive R archive network (CRAN), implements profile reliability [@bulut2013between; @bulut2017], criterion-related profile analysis [@davison_identifying_2002], profile analysis via multidimensional scaling [@pams], moderated profile analysis, profile analysis by group, and a within-person factor model to derive score profiles [@davison2009factor] as well as a variety of graphical methods to visualize profiles. Related profile analysis packages exist in R, such as ``tidyLPA`` [@rosenberg2019tidylpa] and ``mclust`` [@mclust], as well as in Mplus [@muthen2004mplus], however, our approach does not rely on mixture modeling and aims to provide access to a wider range of profile analytic techniques. **An extensive vignette documenting ``profileR`` [@bulut2020] is included with the package as well as vignette to ensure the functionality of the package locally.**

If this is fine with you, then I'll add this to the paper.md and bib file then I'll archive the package on Zendodo.
