### 2020-03-14 c3f84ed
updated paper and vignette naming
