+++ paper.bib
@@ -79,6 +79,14 @@ doi = {10.1080/00273171.2016.1253452}}
   doi = {10.21105/joss.00978}
 }
 
+@article{bulut2020,
+  title={Profile Analysis of Multivariate Data: A Brief Introduction to the profileR Package},
+  author={Bulut, Okan and Desjardins, Christopher David},
+  journal={PsyArXiv},
+  year={2020},
+  doi = {https://doi.org/10.31234/osf.io/sgy8m}
+}
+
 @Article{mclust,
     title = {{mclust} 5: clustering, classification and density estimation using {G}aussian finite mixture models},
     author = {Luca Scrucca and Michael Fop and Thomas Brendan Murphy and Adrian E. Raftery},
