### 2020-03-14 #1941 @cddesja
@arfon, the paper is done. I posted the software on Zendodo.
DOI: 10.5281/zenodo.3711046
Version: 0.3-6
The paper.md and paper.bib files are done.
