### 2020-03-16 #1941 @marcosvital
@cddesja , there is a final invalid DOI. Can you check and correct it?
