+++ paper.bib
@@ -7,8 +7,7 @@
 	journal = {Psychological Methods},
 	author = {Davison, Mark L. and Davenport, Jr., Ernest C.},
 	year = {2002},
-	pages = {468--484},
-	file = {Davison_Davenport_2002_Identifying criterion-related patterns of predictor scores using multiple.pdf:/Users/cdesjardins/Tresors/chris/Zotero/D/Davison_Davenport_2002_Identifying criterion-related patterns of predictor scores using multiple.pdf:application/pdf}
+	pages = {468--484}
 }
 
 @incollection{pams,
@@ -84,7 +83,7 @@ doi = {10.1080/00273171.2016.1253452}}
   author={Bulut, Okan and Desjardins, Christopher David},
   journal={PsyArXiv},
   year={2020},
-  doi = {https://doi.org/10.31234/osf.io/sgy8m}
+  doi = {10.31234/osf.io/sgy8m}
 }
 
 @Article{mclust,
