### 2020-03-16 #1941 @cddesja
Can I look at the final proof just to make it looks okay? Thanks for fixing the DOI, Okan.
Chris

> On Mar 16, 2020, at 9:33 AM, Okan Bulut <notifications@github.com> wrote:
>
> ﻿
> @cddesja @marcosvital , the invalid DOI has been fixed now. Thanks!
>
