### 2020-03-16 #1941 @cddesja
Okay, I changed profileR to ``profileR`` and uploaded the new paper.md.
Everything else looks good to me.
