### 2020-03-17 #1941 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published
