### 2020-03-17 #1941 @okanbulut
> @wjakethompson, @jrosen48 - many thanks for your reviews here and to @marcosvital for editing this submission ✨
>
> @cddesja - your paper is now accepted into JOSS ⚡️🚀💥

Thank you very much!
